#11 skip "Resume game" prompt using built-in variable instead of exe patch
At the end of the loading TppSequence calls TppMain.DisableBlackLoading.
For whatever reason it was not called when player used non-standard
plparts, which caused black screen after loading with everything
else working. Somehow that call (or lack of it) was connected to exe
patch that disabled prompt and tips; replacing that patch fixed the
issue.